Adds the "Toggle Visible" option.
This adds a link to hide/unhide an encrypted element (so the user doesn't have
to see the encrypted data). This is less tested than other releases, but
generally works for me. To specify if element is collapsed or expanded by
default, add the 'COLLAPSED=“1”' entry to the SECRET or LOCK tag.